Restructure Pixy and Ultrasonic for better thread behavior
They are now in separate threads. To get distance data into the vision
calculations for angle, the Pixy class has a member reference to an
Ultrasonic object. Terminating and starting threads has been redone
to be much more intuitive and hopefully crash less.
